Main service refactor - doesn't compile yet
